+++ resources/constants.php
@@ -24,12 +24,12 @@
 
 // App
 
-define('APP_CONFIG_PATH', APP_ROOT_PATH . '/config'); // Path to the application's `config` directory
-define('APP_RESOURCES_PATH', APP_ROOT_PATH . '/resources'); // Path to the application's `resources` directory
-define('APP_STORAGE_PATH', APP_ROOT_PATH . '/storage'); // Path to the application's `storage` directory
+const APP_CONFIG_PATH = APP_ROOT_PATH . '/config'; // Path to the application's `config` directory
+const APP_RESOURCES_PATH = APP_ROOT_PATH . '/resources'; // Path to the application's `resources` directory
+const APP_STORAGE_PATH = APP_ROOT_PATH . '/storage'; // Path to the application's `storage` directory
 
 // Bones
 
 define('BONES_ROOT_PATH', dirname(__FILE__, 2)); // Root path to the Bones directory
-define('BONES_RESOURCES_PATH', BONES_ROOT_PATH . '/resources'); // Path to the Bones `resources` directory
-define('BONES_VERSION', '1.4.0'); // Current Bones version
+const BONES_RESOURCES_PATH = BONES_ROOT_PATH . '/resources'; // Path to the Bones `resources` directory
+const BONES_VERSION = '1.4.1'; // Current Bones version